+++ core/src/plugins/include/python_plugins_common.h
@@ -1,7 +1,7 @@
 /*
    BAREOS® - Backup Archiving REcovery Open Sourced
 
-   Copyright (C) 2020-2020 Bareos GmbH & Co. KG
+   Copyright (C) 2020-2021 Bareos GmbH & Co. KG
 
    This program is Free Software; you can modify it under the terms of
    version three of the GNU Affero General Public License as published by the
@@ -83,19 +83,28 @@
      IO_WRITE = 3
  */
 
-#define ConstSet_StrLong(dict, string, value)           \
-  if (PyDict_SetItem(dict, PyBytes_FromString(#string), \
-                     PyLong_FromLong(value))) {         \
-    return MOD_ERROR_VAL;                               \
-  }                                                     \
-  if (PyModule_AddIntConstant(m, #string, value)) { return MOD_ERROR_VAL; }
-
-#define ConstSet_StrStr(dict, string, value)            \
-  if (PyDict_SetItem(dict, PyBytes_FromString(#string), \
-                     PyBytes_FromString(value))) {      \
-    return MOD_ERROR_VAL;                               \
-  }                                                     \
-  if (PyModule_AddStringConstant(m, #string, value)) { return MOD_ERROR_VAL; }
+#define ConstSet_StrLong(dict, string, value)                                 \
+  {                                                                           \
+    PyObject* stringKey = PyBytes_FromString(#string);                        \
+    PyObject* longValue = PyLong_FromLong(value);                             \
+    if (PyDict_SetItem(dict, stringKey, longValue)) { return MOD_ERROR_VAL; } \
+    if (PyModule_AddIntConstant(m, #string, value)) { return MOD_ERROR_VAL; } \
+    Py_DECREF(stringKey);                                                     \
+    Py_DECREF(longValue);                                                     \
+  }
+#define ConstSet_StrStr(dict, string, value)             \
+  {                                                      \
+    PyObject* stringKey = PyBytes_FromString(#string);   \
+    PyObject* stringValue = PyBytes_FromString(value);   \
+    if (PyDict_SetItem(dict, stringKey, stringValue)) {  \
+      return MOD_ERROR_VAL;                              \
+    }                                                    \
+    if (PyModule_AddStringConstant(m, #string, value)) { \
+      return MOD_ERROR_VAL;                              \
+    }                                                    \
+    Py_DECREF(stringKey);                                \
+    Py_DECREF(stringValue);                              \
+  }
 
 
 /* commonly used definitions in multiple python modules */