+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and since Bareos version 20 this project adheres to [Semantic Versioning](https:
 
 ### Fixed
 - fix possible read/write problems when using droplet with https [PR #773]
+- fix memory leak in python module constants [PR #781]
 
 ### Added
 - systemtests for S3 functionalities (droplet, libcloud) now use https [PR #773]